requirements.txt
setuptools
pytest==7.1.2
pytest-cov==3.0.0
requests
langchain
langchain_community
python-dotenv
mock
gigachain==0.2.6
gigachain_community==0.2.6
gigachain-cli==0.0.25
langgraph==0.2.19
langchain-community==0.2.16
langchain-openai==0.1.23
python-pptx==0.6.23
python-pptx-interface==0.0.12
python-dotenv==1.0.1
duckduckgo-search==6.2.4
bs4
feedparser
streamlit